+++ packages/plugin-calls-ui/src/components/Widget.tsx
@@ -1,12 +1,12 @@
-import { SIP_STATUS_DISCONNECTED, SIP_STATUS_ERROR } from "../lib/enums";
-import { callPropType, sipPropType } from "../lib/types";
+import { SIP_STATUS_DISCONNECTED, SIP_STATUS_ERROR } from '../lib/enums';
+import { callPropType, sipPropType } from '../lib/types';
 
-import Icon from "@erxes/ui/src/components/Icon";
-import Popover from "@erxes/ui/src/components/Popover";
-import React from "react";
-import WidgetPopover from "./WidgetPopover";
-import { WidgetWrapper } from "../styles";
-import { __ } from "@erxes/ui/src/utils";
+import Icon from '@erxes/ui/src/components/Icon';
+import Popover from '@erxes/ui/src/components/Popover';
+import React from 'react';
+import WidgetPopover from './WidgetPopover';
+import { WidgetWrapper } from '../styles';
+import { __ } from '@erxes/ui/src/utils';
 
 type Props = {
   callUserIntegrations: any;
@@ -18,10 +18,11 @@ type Props = {
 
 const Widget = (props: Props, context) => {
   const Sip = context;
-  const isConnected =
+  const isDisConnected =
     !Sip.call ||
     Sip.sip?.status === SIP_STATUS_ERROR ||
-    Sip.sip?.status === SIP_STATUS_DISCONNECTED;
+    Sip.sip?.status === SIP_STATUS_DISCONNECTED ||
+    false;
 
   const onClick = () => {
     const { setHideIncomingCall, hideIncomingCall } = props;
@@ -33,8 +34,8 @@ const Widget = (props: Props, context) => {
   return (
     <Popover
       trigger={
-        <WidgetWrapper $isConnected={isConnected} onClick={onClick}>
-          <Icon icon={isConnected ? "phone-slash" : "phone"} size={23} />
+        <WidgetWrapper $isDisConnected={isDisConnected} onClick={onClick}>
+          <Icon icon={isDisConnected ? 'phone-slash' : 'phone'} size={23} />
         </WidgetWrapper>
       }
       placement="top"