+++ packages/plugin-clientportal-api/src/models/definitions/clientPortal.ts
@@ -63,7 +63,7 @@ export interface IClientPortal {
   tokenExpiration?: number;
   refreshTokenExpiration?: number;
   tokenPassMethod?: 'cookie' | 'header';
-
+  erxesAppToken?: string;
   otpConfig?: IOTPConfig;
   twoFactorConfig?: TwoFactorConfig;
 
@@ -117,12 +117,18 @@ export interface IClientPortal {
   slug?: string;
   template?: string;
   templateId?: string;
-  keywords?: string,
-  copyright?: string
+  keywords?: string;
+  copyright?: string;
 
   externalLinks?: {
     [key: string]: string;
   }
+
+  googleAnalytics?: string;
+  facebookPixel?: string;
+  googleTagManager?: string;
+  vercelProjectId?: string;
+  lastVercelDeploymentId?: string;
 }
 
 interface IStyles {
@@ -366,4 +372,9 @@ export const clientPortalSchema = new Schema({
   keywords: field({ type: String, optional: true }),
   copyright: field({ type: String, optional: true }),
   externalLinks: field({ type: Object, optional: true }),
+  googleAnalytics: field({ type: String, optional: true }),
+  facebookPixel: field({ type: String, optional: true }),
+  googleTagManager: field({ type: String, optional: true }),
+  vercelProjectId: field({ type: String, optional: true }),
+  lastVercelDeploymentId: field({ type: String, optional: true }),
 });