+++ packages/plugin-payment-api/src/api/ErxesPayment.ts
@@ -11,6 +11,8 @@ import { ITransactionDocument } from '../models/definitions/transactions';
 import { MinuPayAPI } from './minupay/api';
 import { GolomtAPI } from './golomt/api';
 import { StripeAPI } from './stripe/api';
+import { KhanbankAPI } from './khanbank/api';
+import { getEnv } from '@erxes/api-utils/src/core';
 
 class ErxesPayment {
   public socialpay!: SocialPayAPI;
@@ -24,50 +26,58 @@ class ErxesPayment {
   public minupay!: MinuPayAPI;
   public golomt!: GolomtAPI;
   public stripe!: StripeAPI;
+  public khanbank!: KhanbankAPI;
   public domain: string;
 
   private payment: any;
 
-  constructor(payment: IPaymentDocument, domain?: string) {
+  constructor(payment: IPaymentDocument, subdomain?: string) {
     this.payment = payment;
-    this.domain = domain || '';
+    const DOMAIN = getEnv({ name: 'DOMAIN' })
+      ? `${getEnv({ name: 'DOMAIN' })}/gateway`
+      : 'http://localhost:4000';
+    const apiDomain = DOMAIN.replace('<subdomain>', subdomain || '');
+    this.domain = apiDomain;
 
     switch (payment.kind) {
       case 'socialpay':
         this.socialpay = new SocialPayAPI(payment.config);
         break;
       case 'storepay':
-        this.storepay = new StorePayAPI(payment.config, domain);
+        this.storepay = new StorePayAPI(payment.config, this.domain);
         break;
       case 'qpay':
         this.qpay = new QpayAPI(
           { ...payment.config, branchCode: payment.name },
-          domain
+          this.domain
         );
         break;
       case 'monpay':
-        this.monpay = new MonpayAPI(payment.config, domain);
+        this.monpay = new MonpayAPI(payment.config, this.domain);
         break;
       case 'paypal':
         this.paypal = new PaypalAPI(payment.config);
         break;
       case 'wechatpay':
-        this.wechatpay = new WechatPayAPI(payment.config, domain);
+        this.wechatpay = new WechatPayAPI(payment.config, this.domain);
         break;
       case 'qpayQuickqr':
-        this.qpayQuickqr = new QPayQuickQrAPI(payment.config, domain);
+        this.qpayQuickqr = new QPayQuickQrAPI(payment.config, this.domain);
         break;
       case 'pocket':
-        this.pocket = new PocketAPI(payment.config, domain);
+        this.pocket = new PocketAPI(payment.config, this.domain);
         break;
       case 'minupay':
-        this.minupay = new MinuPayAPI(payment.config, domain);
+        this.minupay = new MinuPayAPI(payment.config, this.domain);
         break;
       case 'golomt':
-        this.golomt = new GolomtAPI(payment.config, domain);
+        this.golomt = new GolomtAPI(payment.config, this.domain);
         break;
       case 'stripe':
-        this.stripe = new StripeAPI(payment.config, domain);
+        this.stripe = new StripeAPI(payment.config, this.domain);
+        break;
+      case 'khanbank':
+        this.khanbank = new KhanbankAPI(payment.config, subdomain || '');
         break;
       default:
         break;